requirements.txt
streamlit
python-dotenv
google-generativeai
youtube_transcript_api
reportlab
gtts
google-api-python-client




